enhancement: add support to configure channel timeout in pyFG.FortiOS
Add a keyword argument 'chan_timeout' to pyFG.FortiOS.__init__, to allow
users to configure channel timeout, a timeout to wait for responces from
devices via ssh 'channel'.

It was set to 5 [sec] in pyFG.FortiOS.execute_command with using
paramiko.Channel.settimeout, however it's too short time for some
environments and this change is necessary to allow to adjust that value
as needed for such environments.